// src/containers/Dice/state/interfaces.ts
export enum Rolls {
  ONE = 1,
  TWO = 2,
  THREE = 3,
  FOUR = 4,
  FIVE = 5,
  SIX = 6,
}

export interface IState {
  roll: Rolls;
  isDieRollAllowed: boolean;
  isDieRollValid: boolean;
  isTimerRunning: boolean,
  TIME_LIMIT: number,
  timePassed: number,
  timeLeft: number,
  remainingPathColor: string,
}